% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covers.R
\name{cover}
\alias{cover}
\title{Covers / Lottery}
\usage{
cover(pick = 3, pool = 10, replace = TRUE, ordered = TRUE, powerBall = NA)
}
\arguments{
\item{pick}{the number of numbers drawn}

\item{pool}{the pool size}

\item{replace}{TRUE if a number can be drawn more than once}

\item{ordered}{TRUE if the order numbers are drawn/picked affects winning}

\item{powerBall}{NA for non-powerball games, otherwise requires the size
of the powerball pool}
}
\value{
returns the numbers drawn
}
\description{
Provides a basic framework to define a lottery type experiment
}
\details{
options that may need to be implemented:
replacement, in order requirements and powerball
an additional function to check results should be implemented
after the framework is in place.

Still need to add some method of specifying a game starts at 0 instead of 1
as is common with pick 2/3/4/5 games.
}
\examples{
cover(pick = 5, pool = 69, replace = FALSE, ordered = FALSE, powerBall = 26)
cover(pick = 5, pool = 36, replace = FALSE, ordered = FALSE, powerBall = NA)
}
