% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{flipCoin}
\alias{flipCoin}
\title{Simulates a coin flip}
\usage{
flipCoin(nTimes = 1)
}
\arguments{
\item{nTimes}{the number of times to flip}
}
\value{
returns the results as a list
}
\description{
Simulates flipping a single (fair) coin some number of times
}
\examples{
flipCoin( nTimes = 1 )
}
