% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollDice.R
\name{rollDice}
\alias{rollDice}
\title{Rolls Dice}
\usage{
rollDice(nSides = 6, nDice = 1, nTimes = 1, sum = FALSE)
}
\arguments{
\item{nSides}{the number of sides each die should have}

\item{nDice}{the number of dice to roll each time}

\item{nTimes}{the number of times to roll them}

\item{sum}{if TRUE then returns the sum of the dice instead of the individual
values}
}
\value{
returns the results (individual rolls, or sums) as a list
}
\description{
Repeatedly rolls any number of identical, fair dice with configurable sides
}
\examples{
rollDice(nSides = 6, nDice = 3, nTimes = 2, sum = FALSE)
}
