% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosencrantzEquiv.R
\name{rosencrantzEquiv}
\alias{rosencrantzEquiv}
\title{Equivalent Odds in Coin Flips}
\usage{
rosencrantzEquiv(oneIn = 1)
}
\arguments{
\item{oneIn}{the odds to translate into coin flips}
}
\value{
returns the equivalent number of repeated coin flips
}
\description{
Determine how many times in a row you would need a coin flip to end the
same way to be equivalent to the odds in question.
}
\details{
e.g., 1 in 1 Million is roughly the same as flipping a coin and having
it land on heads 20 times in a row.
}
\examples{
 rosencrantzEquiv(oneIn = 1000000)
}
