% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosencrantzTest.R
\name{rosencrantzTest}
\alias{rosencrantzTest}
\title{Simulates a Rosencrantz Experiment}
\usage{
rosencrantzTest(rNumber = 1, seed = NA, upTo = FALSE)
}
\arguments{
\item{rNumber}{the number of repeated outcomes to stop at}

\item{seed}{a random seed, for repeatable results, otherwise the function
will initialize the random number generator with the system time}

\item{upTo}{if TRUE the function will perform and return the results of all
tests from 1 up to the rNumber. Useful for demonstrations and plotting.}
}
\description{
Simulates flipping a coin until it has come up with the same result
some specified number of times in a row, and returns how many attempts
it took.
}
\examples{
rosencrantzTest( rNumber = 5, seed = 1, upTo = FALSE )
}
